backend/requirements.txt
fastapi>=0.100.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyjwt>=2.8.0
uvicorn[standard]>=0.24.0
